Modified repositories

+++ src/main/java/srl/narrel/demo/repositories/CondominioRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import srl.narrel.demo.models.CondominioModel;
 
-public interface CondominioRepository extends JpaRepository<CondominioModel, Long> {
+public interface CondominioRepository extends JpaRepository<CondominioModel, Integer> {
  
 }
